feat: integrate react-redux for state management and refactor components to use Redux
