+++ part6/redux-notes/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
+    "react-redux": "^9.2.0",
     "redux": "^5.0.1"
   },
   "devDependencies": {
