+++ part6/redux-notes/src/components/NewNote.jsx
@@ -1,20 +1,21 @@
-import { store, createNote } from "../reducers/noteReducer";
+import { createNote } from "../reducers/noteReducer";
+import { useDispatch } from "react-redux";
 
 const NewNote = () => {
+  const dispatch = useDispatch();
+
   const addNote = (event) => {
     event.preventDefault();
     const content = event.target.note.value;
     event.target.note.value = "";
-    store.dispatch(createNote(content));
+    dispatch(createNote(content));
   };
 
   return (
-    <div>
       <form onSubmit={addNote}>
         <input name="note" />
         <button type="submit">add</button>
       </form>
-    </div>
   );
 };
